@@ -20,7 +20,7 @@ module.exports = function(config) {
       'node_modules/zone.js/dist/async-test.js',
       'node_modules/zone.js/dist/fake-async-test.js',
 
-      'node_modules/rxjs/bundles/Rx.{js,map}',
+      'node_modules/rxjs/bundles/rxjs.umd.{js,map}',
 
       ...getAngularFiles(['core','common','compiler','platform-browser','platform-browser-dynamic']),
 
@@ -32,7 +32,6 @@ module.exports = function(config) {
       'dist/packages-dist/bundles/database.umd.{js,map}',
       'dist/packages-dist/bundles/firestore.umd.{js,map}',
       'dist/packages-dist/bundles/storage.umd.{js,map}',
-      'dist/packages-dist/bundles/database-deprecated.umd.{js,map}',
       'dist/packages-dist/bundles/test.umd.{js,map}',
     ],
 
