@@ -1,7 +1,6 @@
 import { SettableMetadata, UploadMetadata, Reference, StringFormat } from '@firebase/storage-types';
 import { createUploadTask, AngularFireUploadTask } from './task';
-import { Observable } from 'rxjs';
-import { from } from 'rxjs/observable/from';
+import { Observable, from } from 'rxjs';
 import { FirebaseZoneScheduler } from 'angularfire2';
 
 export interface AngularFireStorageReference {
@@ -31,9 +30,9 @@ export function createStorageRef(ref: Reference, scheduler: FirebaseZoneSchedule
         from(ref.getMetadata())
       )
     ),
-    delete: () => Observable.from(ref.delete()),
+    delete: () => from(ref.delete()),
     child: (path: string) => createStorageRef(ref.child(path), scheduler),
-    updateMetatdata: (meta: SettableMetadata) => Observable.from(ref.updateMetadata(meta)),
+    updateMetatdata: (meta: SettableMetadata) => from(ref.updateMetadata(meta)),
     put: (data: any, metadata?: UploadMetadata) => {
       const task = ref.put(data, metadata);
       return createUploadTask(task);
